refactor(ecs): split service.rs into 7 resource impl files
service.rs grew to 3291 lines after Phase I extraction. Already had
multiple impl EcsService blocks; extract each non-setup block to a
sibling file via #[path] modules:

- service_clusters.rs (265 lines): create/describe/list/update/delete_cluster.
- service_task_definitions.rs (363 lines): register/describe/deregister/list_task_definition(s/families).
- service_tagging.rs (236 lines): tag_resource, untag_resource, list_tags_for_resource.
- service_account_settings.rs (155 lines): account_setting CRUD.
- service_tasks.rs (397 lines): run_task, start_task, stop_task, describe/list_tasks, exec.
- service_services_resource.rs (603 lines): create/update/delete/describe/list_service.
- service_container_instances_etc.rs (1072 lines): container_instance ops, attributes, capacity_providers, task_sets.

service.rs: 3291 -> 296 lines (-91%).
